Great Improvements, fixes and more!
Fixes for RMB shortcut
New brush texture preview, toggle it on in the preferences dropdown menu (gear of the tool header)
Preview of the brand new reference system, this thing is still in progress and still need so much work and time to finish it up
That was hard and late BUT...! Now you can edit the dyntopo values!! Also they are saved into the preferences, also if you have autosave turned on you dont have to worry about it. Values can be changed in preferences and in the dropdown menu of dyntopo stages over the tool header.
Now in the brush dropdown menu (collapsed only) in the tool header you can select if you want to render your custom icon with alpha or without, and select a color for the background of this icon if you're not using alpha, this way you can organize brushes per colors to have them better locatable.
Now all ui elements have description to know what they do!
Others fixes :)