Better support with Remeshers. Multi-platform support.
Instant Meshes is included and supported for all platforms.
Quadriflow is included for Windows users by default. Linux and MacOS is not included but you can download the source code and get the compiling instructions with a new button on the addon preferences over the preferences windows.
Recommendation for Windows Users: to update the included quadriflow program